Minor edits to organization
