+++ app/__init__.py
@@ -1,13 +1,9 @@
 from flask import Flask
 
-# Declaring and registering the views
-from app.full_time_staff import FullTimeStaffView
-from app.student_employees import StudentEmployeesView
-from app.views import View
+# Declaring and registering the view
+from app.views.__init__ import ShiftsView
 
 app = Flask(__name__)
 app.config.from_object('config')
 
-View.register(app)
-FullTimeStaffView.register(app)
-StudentEmployeesView.register(app)
+ShiftsView.register(app, route_base='/')
