+++ app/static/shift_buttons.js
@@ -4,7 +4,8 @@ $(document).ready(function() {
     $("#generate-shifts").click(function() {
        let input_data = {'clicked': 'true'};
        $.post('/generate_shifts', input_data, function(output_data) {
-            alert('CLICKED');
+            // do something
        });
+       alert('CLICK')
     });
 });
