--- app/templates/full_time_staff/generate_shifts.html
@@ -1,5 +1,0 @@
-{% extends "full_time_staff/index.html" %}
-
-{% block tab_content %}
-    <!-- Test code for how to load shift information, this is incomplete -->
-{% endblock %}
