+++ app/templates/full_time_staff/index.html
@@ -1,27 +1,16 @@
 {% extends 'help_desk_base.html' %}
 
-{% block header %}
-{% endblock %}
-
 {% block body_content %}
     <!-- Test code / incomplete -->
-    <a class="nav-buttons btn btn-link" href="https://tinyurl.com/y6tp7ylg">Access full spreadsheet</a>
-    <p>
-      <a class="btn" id="show-shifts" data-toggle="collapse" href="#collapse-sheet" role="button" aria-expanded="false" aria-controls="collapse-sheet">
-        Show flagged shifts
-      </a>
-    </p>
-    <div class="collapse" id="collapse-sheet">
-      <div class="card card-body">
-       <iframe style="width: 100%; height: 30rem;" src="https://docs.google.com/a/bethel.edu/spreadsheets/d/e/2PACX-1vQ_JcEenZkbJw59E-FH94YHiNyQOlzm6AcIhj8XSwlV6G9OcLruQWhqZyu4I1DI5erNJwz8s8Bgbx-X/pubhtml?gid=1314546665&amp;single=true&amp;widget=true&amp;headers=false"></iframe>
-      </div>
+    <div class="sub-nav">
+        <a class="nav-button btn btn-success" id="generate-shifts" href="{{ url_for('ShiftsView:generate_shifts') }}">Generate Flagged Shifts</a>
+        <br/><br/>
+        <a class="nav-buttons btn btn-info" href="https://tinyurl.com/y6tp7ylg">Access full spreadsheet</a>
+        <br/><br/>
+        <a class="nav-buttons btn btn-primary" id="show-shifts" href="{{ url_for('ShiftsView:show_shifts') }}">Show Shifts</a>
     </div>
     {% block tab_content %}
 
     {% endblock %}
 
 {% endblock %}
-
-<!--
-<iframe style="width: 100%; height: 30rem;" src="https://docs.google.com/a/bethel.edu/spreadsheets/d/e/2PACX-1vQ_JcEenZkbJw59E-FH94YHiNyQOlzm6AcIhj8XSwlV6G9OcLruQWhqZyu4I1DI5erNJwz8s8Bgbx-X/pubhtml?gid=1314546665&amp;single=true&amp;widget=true&amp;headers=false"></iframe>
--->
