+++ app/templates/full_time_staff/show_shifts.html
@@ -1,0 +1,6 @@
+{% extends "full_time_staff/index.html" %}
+
+{% block tab_content %}
+    <!-- Test code for how to load shift information, this is incomplete -->
+    <iframe style="width: 100%; height: 30rem;" src="https://docs.google.com/a/bethel.edu/spreadsheets/d/e/2PACX-1vQ_JcEenZkbJw59E-FH94YHiNyQOlzm6AcIhj8XSwlV6G9OcLruQWhqZyu4I1DI5erNJwz8s8Bgbx-X/pubhtml?gid=1314546665&amp;single=true&amp;widget=true&amp;headers=false"></iframe>
+{% endblock %}
