+++ app/templates/student_employees/index.html
@@ -30,6 +30,8 @@
             </div>
         </div>
         <div class="row">
+            <h5>What I want alerts to look like after clicking a button and reading whether the time clock
+            was successful</h5>
             <div class="alert alert-success text-center m-3 mb-3 mx-auto" role="alert" style="width: 30rem;">
                 <h4 class="alert-heading">Sign in/out successful</h4>
             </div>
